flarevm_cli/requirements.txt
anthropic>=0.40.0
requests>=2.32.3
python-dotenv>=1.0.0

